# [asset-reconciliation] Improve performance for never materialized root partitions