--- a/python_modules/dagster/dagster/_core/definitions/asset_reconciliation_sensor.py
+++ b/python_modules/dagster/dagster/_core/definitions/asset_reconciliation_sensor.py
@@ -27,7 +27,6 @@ from dagster._core.definitions.freshness_policy import FreshnessConstraint
 from dagster._core.definitions.time_window_partitions import (
     TimeWindow,
     TimeWindowPartitionsDefinition,
-    TimeWindowPartitionsSubset,
 )
 from dagster._core.storage.tags import PARTITION_NAME_TAG
 from dagster._utils.caching_instance_queryer import CachingInstanceQueryer
@@ -68,30 +67,27 @@ class AssetReconciliationCursor(NamedTuple):
         self, asset_key: AssetKey, asset_graph
     ) -> Iterable[str]:
         partitions_def = asset_graph.get_partitions_def(asset_key)
+
+        materialized_or_requested_subset = (
+            self.materialized_or_requested_root_partitions_by_asset_key.get(
+                asset_key, partitions_def.empty_subset()
+            )
+        )
+
         if isinstance(partitions_def, TimeWindowPartitionsDefinition):
-            last_window = partitions_def.get_last_partition_window()
-            default_subset = TimeWindowPartitionsSubset(
-                partitions_def=partitions_def,
-                included_time_windows=[
-                    TimeWindow(partitions_def.start, last_window.end - datetime.timedelta(days=1))
-                ],
-                num_partitions=0,
-            )
-            keys = partitions_def.get_partition_keys_in_range(
-                partitions_def.get_partition_key_range_for_time_window(
-                    TimeWindow(
-                        start=last_window.start - datetime.timedelta(days=1), end=last_window.end
-                    )
+            allowable_time_window = allowable_time_window_for_partitions_def(partitions_def)
+            if allowable_time_window is None:
+                return []
+            # for performance, only iterate over keys within the allowable time window
+            return [
+                partition_key
+                for partition_key in partitions_def.get_partition_keys_in_time_window(
+                    allowable_time_window
                 )
-            )
-            return keys
-            print(last_window)
-            print(default_subset)
+                if partition_key not in materialized_or_requested_subset
+            ]
         else:
-            default_subset = partitions_def.empty_subset()
-        return self.materialized_or_requested_root_partitions_by_asset_key.get(
-            asset_key, default_subset
-        ).get_partition_keys_not_in_subset()
+            return materialized_or_requested_subset.get_partition_keys_not_in_subset()
 
     def with_updates(
         self,
@@ -291,6 +287,21 @@ def find_never_materialized_or_requested_root_asset_partitions(
     )
 
 
+def allowable_time_window_for_partitions_def(
+    partitions_def: TimeWindowPartitionsDefinition,
+) -> Optional[TimeWindow]:
+    """Returns a time window encompassing the partitions that the reconciliation sensor is currently
+    allowed to materialize for this partitions_def
+    """
+    latest_partition_window = partitions_def.get_last_partition_window()
+    if latest_partition_window is None:
+        return None
+    return TimeWindow(
+        start=latest_partition_window.start - datetime.timedelta(days=1),
+        end=latest_partition_window.end,
+    )
+
+
 def candidates_unit_within_allowable_time_window(
     asset_graph: AssetGraph, candidates_unit: Iterable[AssetKeyPartitionKey]
 ):
@@ -308,14 +319,15 @@ def candidates_unit_within_allowable_time_window(
 
     partitions_def = cast(TimeWindowPartitionsDefinition, partitions_def)
 
-    latest_partition_window = partitions_def.get_last_partition_window()
-    if latest_partition_window is None:
+    allowable_time_window = allowable_time_window_for_partitions_def(partitions_def)
+    if allowable_time_window is None:
         return False
 
     candidate_partition_window = partitions_def.time_window_for_partition_key(partition_key)
-    time_delta = latest_partition_window.end - candidate_partition_window.end
-
-    return time_delta < datetime.timedelta(days=1)
+    return (
+        candidate_partition_window.start > allowable_time_window.start
+        and candidate_partition_window.end <= allowable_time_window.end
+    )
 
 
 def determine_asset_partitions_to_reconcile(