### Summary & Motivation

A followup from the discussion here: https://github.com/dagster-io/dagster/pull/11582. This dramatically improves performance.

#### Test Setup

A 30-asset multi-asset, downstream of two 15-minute partitioned assets (start-date 2022-01-01). 

#### Benchmarks

I originally tested this before Claire's performance improvements to TimeWindowPartitionsSubset were merged in. In that state, it took 114 seconds (!!) to run my test, the majority of it spent iterating over time windows.

Claire's changes made a big difference, cutting down the total time to 28 seconds, although the performance scaled pretty poorly as the number of source assets increased (i.e. if you had 5 root assets instead of 2, it'd go back to taking over a minute).

After these changes, the original test completes in under 4 seconds, and with 100 source assets, the total time was around 17 seconds, with only around 25% of that time spent within this codepath.

### How I Tested These Changes